requirements.txt
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
gradio>=5.0.0
pdfplumber>=0.11.0
transformers>=4.40.0
torch>=2.3.0